refactor dsa out of run utils
